SGBG-19 refactor: 모든 경로를 절대경로로 변경 및 컴포넌트 파일 위치 변경
