+++ src/components/base/Textarea/index.tsx
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
 
-import { Flex } from '../Flex';
-import { Txt } from '../Txt';
+import { Flex } from '@/components/base/Flex';
+import { Txt } from '@/components/base/Txt';
 
 import { textareaStyle } from '@/styles/base/input.css';
 import { sprinkles } from '@/styles/sprinkles.css';
