+++ src/components/common/Layout/MainLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 
-import { Footer } from '../Footer';
+import { Footer } from '@/components/common/Footer';
 
 export const MainLayout = () => {
   return (
