+++ src/components/onboarding/BasicInfo/index.tsx
@@ -1,17 +1,16 @@
 import { useFormContext, useWatch } from 'react-hook-form';
 
+import { Button } from '@/components/base/Button';
+import { Flex } from '@/components/base/Flex';
+import { Selector } from '@/components/base/Selector';
+import { Spacing } from '@/components/base/Spacing';
+import { Txt } from '@/components/base/Txt';
+import { Question } from '@/components/onboarding/Question';
+import { StepBar } from '@/components/onboarding/StepBar';
+import { Title } from '@/components/onboarding/Title';
 import { ageList, genderList, STEPS } from '@/constants/onboarding';
 import type { AgeListType, GenderListType } from '@/types/onboarding';
 import { handleSelector } from '@/utils/home/handleSelector';
-
-import { Button } from '../base/Button';
-import { Flex } from '../base/Flex';
-import { Selector } from '../base/Selector';
-import { Spacing } from '../base/Spacing';
-import { Txt } from '../base/Txt';
-import { Question } from './Question';
-import { StepBar } from './StepBar';
-import { Title } from './Title';
 
 type FormValues = {
   gender: GenderListType | '';
