+++ src/components/onboarding/Genre/index.tsx
@@ -1,16 +1,15 @@
 import { useFormContext, useWatch } from 'react-hook-form';
 
+import { Button } from '@/components/base/Button';
+import { Flex } from '@/components/base/Flex';
+import { MultiSelector } from '@/components/base/MultiSelector';
+import { Spacing } from '@/components/base/Spacing';
+import { Txt } from '@/components/base/Txt';
+import { StepBar } from '@/components/onboarding/StepBar';
+import { Title } from '@/components/onboarding/Title';
 import { genreList, STEPS } from '@/constants/onboarding';
 import type { GenreListType } from '@/types/onboarding';
 import { handleMultiSelector } from '@/utils/home/handleMultiSelector';
-
-import { Button } from '../base/Button';
-import { Flex } from '../base/Flex';
-import { MultiSelector } from '../base/MultiSelector';
-import { Spacing } from '../base/Spacing';
-import { Txt } from '../base/Txt';
-import { StepBar } from './StepBar';
-import { Title } from './Title';
 
 type FormValues = {
   genre: GenreListType[];
