+++ src/main.tsx
@@ -2,10 +2,10 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { RouterProvider } from 'react-router-dom';
 
-import { router } from './config/router.tsx';
+import { router } from '@/config/router.tsx';
 
-import './styles/theme.css.ts';
-import './styles/reset.css.ts';
+import '@/styles/theme.css.ts';
+import '@/styles/reset.css.ts';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
